Add files via upload
Using python, mysql and json to create a database with a google map api  -  First we run the geoload.py, retrieving the adresses, and after that the geodump.py,  where we run all the locations connected to where.html. If we open where.html, we see all the marked locations that were stored on our sqlite database.